Feature: Cvt Temperature Done

+++ Calculator.java
@@ -1,19 +1,19 @@
 package MergePractice;
 public class Calculator {
     // Add two numbers
-    public static int add(int a, int b) {
+    public static double add(double a, double b) {
         return a + b;
     }
     // Subtract two numbers
-    public static int subtract(int a, int b) {
+    public static double subtract(double a, double b) {
         return a - b;
     }
     // Multiply two numbers
-    public static int multiply(int a, int b) {
+    public static double multiply(double a, double b) {
         return a * b;
     }
     // Divide two numbers
-    public static int divide(int a, int b) {
+    public static double divide(double a, double b) {
         return a / b;
     }
 }
